backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0

# Firebase (includes Firestore)
firebase-admin==6.4.0
google-cloud-firestore==2.14.0

# AI
anthropic==0.18.1

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# File processing
python-docx==1.1.0
PyPDF2==3.0.1
pdfplumber==0.10.3
reportlab==4.0.8
Pillow==10.2.0

# Async tasks (optional - for background processing)
celery==5.3.6
redis==5.0.1

# Utils
httpx>=0.24,<0.26
python-dateutil==2.8.2
tenacity==8.2.3
structlog==24.1.0
aiofiles==23.2.1
markdown==3.5.2
